// src/communication/models/put-user-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * QuizApi
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface PutUserRequest
 */
export interface PutUserRequest {
    /**
     * 
     * @type {string}
     * @memberof PutUserRequest
     */
    username: any;
    /**
     * 
     * @type {string}
     * @memberof PutUserRequest
     */
    email: any;
    /**
     * 
     * @type {string}
     * @memberof PutUserRequest
     */
    avatar?: any | null;
    /**
     * 
     * @type {string}
     * @memberof PutUserRequest
     */
    password?: any | null;
    /**
     * 
     * @type {string}
     * @memberof PutUserRequest
     */
    oldPassword?: any | null;
}